Binary search implementation.
